requirements.txt
django
djangorestframework
python-slugify
djoser
unidecode
pillow
pdfkit